mappers + dto added, some dependencies was added, project structure refactored.
