+++ src/main/java/ru/raiffeisen/ipr/Application.java
@@ -1,37 +1,36 @@
-package application;
+package ru.raiffeisen.ipr;
 
-import entity.Client;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.stereotype.Component;
-import repository.ClientRepository;
-import service.ClientService;
-import service.PlanService;
-import service.impl.ClientServiceImpl;
+import ru.raiffeisen.ipr.entity.Client;
+import ru.raiffeisen.ipr.service.ClientService;
+import ru.raiffeisen.ipr.service.PlanService;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
 
 @SpringBootApplication
-@EnableJpaRepositories(basePackages = {"repository"})
-@EntityScan(basePackages = {"entity"})
-@ComponentScan(basePackages = "service")
+@EnableJpaRepositories(basePackages = {"ru.raiffeisen.ipr.repository"})
+@EntityScan(basePackages = {"ru.raiffeisen.ipr.entity"})
+//@ComponentScan(basePackages = "ru.raiffeisen.ipr")
 public class Application {
     public static void main(String[] args)throws ParseException {
         ApplicationContext context = SpringApplication.run(Application.class, args);
 
         /**-------- Operation with ClientService --------**/
 
-//        ClientService clientService = context.getBean(ClientService.class);
+        ClientService clientService = context.getBean(ClientService.class);
 
         /** CreateClient function **/
 //        clientService.insertClient("Mark", "Bryzgalov","marik33409@mail.ru", false);
@@ -101,5 +100,8 @@ public class Application {
 
         /** EditTaskList function **/
 //        planService.updateTaskList(false, "Заствить Марка работать", 1l);
+
+        /** TestTestTest **/
+
     }
 }
