+++ src/main/java/ru/raiffeisen/ipr/repository/ClientRepository.java
@@ -1,19 +1,12 @@
-package repository;
+package ru.raiffeisen.ipr.repository;
 
-import entity.Client;
-import org.hibernate.annotations.SQLInsert;
+import ru.raiffeisen.ipr.entity.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.CrudRepositoryExtensionsKt;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 @Repository
 public interface ClientRepository extends JpaRepository<Client, Long> {
