+++ src/main/java/ru/raiffeisen/ipr/repository/MembersListRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package ru.raiffeisen.ipr.repository;
 
-import entity.MembersList;
+import ru.raiffeisen.ipr.entity.MembersList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MembersListRepository extends JpaRepository<MembersList, Integer> {
