+++ src/main/java/ru/raiffeisen/ipr/repository/PlanRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package ru.raiffeisen.ipr.repository;
 
-import entity.Plan;
+import ru.raiffeisen.ipr.entity.Plan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
