+++ src/main/java/ru/raiffeisen/ipr/repository/TasksListRepository.java
@@ -1,14 +1,12 @@
-package repository;
+package ru.raiffeisen.ipr.repository;
 
-import entity.TasksList;
+import ru.raiffeisen.ipr.entity.TasksList;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.sql.Date;
 
 @Repository
 public interface TasksListRepository extends JpaRepository<TasksList, Long> {
